+++ coverage-tests/src/test/java/coverage/GlobalSetup.java
@@ -3,6 +3,8 @@ package coverage;
 import com.applitools.connectivity.RestClient;
 import com.applitools.connectivity.ServerConnector;
 import com.applitools.eyes.*;
+import coverage.exceptions.MissingEnvVarException;
+import org.openqa.selenium.WebDriver;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeSuite;
 
@@ -10,11 +12,19 @@ import javax.ws.rs.HttpMethod;
 import javax.ws.rs.core.UriBuilder;
 import java.net.URI;
 
-public class GlobalSetup extends BuildDriver {
+public class GlobalSetup {
 
+    protected WebDriver driver;
+    protected WebDriver eyesDriver;
     protected static BatchInfo batch;
     protected static String apiKey;
+    public static boolean useDocker;
+    public static boolean CI;
+    public static String EG_URL;
 
+    public WebDriver getDriver() {
+        return eyesDriver == null ? driver : eyesDriver;
+    }
     @BeforeSuite
     public void globalSetup() {
         String name = System.getenv("APPLITOOLS_BATCH_NAME");
@@ -23,6 +33,24 @@ public class GlobalSetup extends BuildDriver {
         String id = System.getenv("APPLITOOLS_BATCH_ID");
         if (id != null) batch.setId(id);
         apiKey = System.getenv("APPLITOOLS_API_KEY");
+        EG_URL = System.getenv("EXECUTION_GRID_URL");
+        String CI = System.getenv("CI");
+        GlobalSetup.CI = CI != null && CI.equals("true");
+        useDocker = !GlobalSetup.CI;
+        String envVar = System.getenv("USE_SELENIUM_DOCKER");
+        if (envVar != null && envVar.equals("false")) {
+            useDocker = false;
+        }
+        if(!useDocker) {
+            String chromeDriverPath = System.getenv("CHROME_DRIVER_PATH");
+            if(chromeDriverPath == null) throw new MissingEnvVarException("CHROME_DRIVER_PATH");
+            System.setProperty("webdriver.chrome.driver", chromeDriverPath);
+            String firefoxDriverPath = System.getenv("FIREFOX_DRIVER_PATH");
+            if(firefoxDriverPath == null) throw new MissingEnvVarException("FIREFOX_DRIVER_PATH");
+            System.setProperty("webdriver.gecko.driver", firefoxDriverPath);
+        }
+
+
     }
 
     @AfterSuite