+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/AndroidScrollPositionProvider.java
@@ -173,11 +173,6 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
         curScrollPos = getScrollPosFromScrollData(contentSize, lastScrollData, supposedScrollAmt, isDown);
     }
 
-    public Location scrollDown(boolean returnAbsoluteLocation) {
-        scroll(true);
-        return getCurrentPositionWithoutStatusBar(returnAbsoluteLocation);
-    }
-
     @Override
     public void scrollTo(int startX, int startY, int endX, int endY, boolean shouldCancel) {
         logger.log(TraceLevel.Debug, eyesDriver.getTestId(), Stage.CHECK,
@@ -209,6 +204,7 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
                 hiddenElement.click();
                 scrolled = true;
                 try { Thread.sleep(1000); } catch (InterruptedException ignored) {}
+                hiddenElement.clear();
             }
         } catch (NoSuchElementException | StaleElementReferenceException e) {
             GeneralUtils.logExceptionStackTrace(logger, Stage.CHECK, e);
@@ -225,8 +221,42 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
                 hiddenElement.click();
                 scrolled = true;
                 try { Thread.sleep(1000); } catch (InterruptedException ignored) {}
+                hiddenElement.clear();
             }
         } catch (NoSuchElementException | StaleElementReferenceException e) {
+            GeneralUtils.logExceptionStackTrace(logger, Stage.CHECK, e);
+        }
+        return scrolled;
+    }
+
+    public int getBehaviorOffsetWithHelperLibrary(String elementId) {
+        int offset = -1;
+        try {
+            MobileElement hiddenElement = ((AndroidDriver<AndroidElement>) driver).findElement(MobileBy.AndroidUIAutomator("new UiSelector().description(\"EyesAppiumHelperEDT\")"));
+            if (hiddenElement != null) {
+                hiddenElement.setValue("behaviorOffset;"+elementId+";0;0");
+                hiddenElement.click();
+                offset = Integer.parseInt(hiddenElement.getText());
+                hiddenElement.clear();
+            }
+        } catch (NoSuchElementException | NumberFormatException | StaleElementReferenceException e) {
+            GeneralUtils.logExceptionStackTrace(logger, Stage.CHECK, e);
+        }
+        return offset;
+    }
+
+    public boolean tryScrollBehaviorOffsetWithHelperLibrary(String elementId, int offset) {
+        boolean scrolled = false;
+        try {
+            MobileElement hiddenElement = ((AndroidDriver<AndroidElement>) driver).findElement(MobileBy.AndroidUIAutomator("new UiSelector().description(\"EyesAppiumHelperEDT\")"));
+            if (hiddenElement != null) {
+                hiddenElement.setValue("behaviorScroll;"+elementId+";" + offset + ";0");
+                hiddenElement.click();
+                scrolled = true;
+                try { Thread.sleep(1500); } catch (InterruptedException ignored) {}
+                hiddenElement.clear();
+            }
+        } catch (NoSuchElementException | NumberFormatException | StaleElementReferenceException e) {
             GeneralUtils.logExceptionStackTrace(logger, Stage.CHECK, e);
         }
         return scrolled;
@@ -251,7 +281,8 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
                     element.getAttribute("className").equals("androidx.recyclerview.widget.RecyclerView") ||
                     element.getAttribute("className").equals("androidx.viewpager2.widget.ViewPager2") ||
                     element.getAttribute("className").equals("android.widget.ListView") ||
-                    element.getAttribute("className").equals("android.widget.GridView")) {
+                    element.getAttribute("className").equals("android.widget.GridView") ||
+                    element.getAttribute("className").equals("android.widget.ScrollView")) {
                 try {
                     String scrollableContentSize = getScrollableContentSize(element.getAttribute("resourceId"));
                     try {
@@ -350,7 +381,8 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
                     className.equals("androidx.recyclerview.widget.RecyclerView") ||
                     className.equals("androidx.viewpager2.widget.ViewPager2") ||
                     className.equals("android.widget.ListView") ||
-                    className.equals("android.widget.GridView")) {
+                    className.equals("android.widget.GridView") ||
+                    className.equals("android.widget.ScrollView")) {
                 try {
                     String scrollableContentSize = getScrollableContentSize(activeScroll.getAttribute("resourceId"));
                     try {
@@ -366,6 +398,14 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
                     }
                     GeneralUtils.logExceptionStackTrace(logger, Stage.CHECK, e);
                 }
+                int position;
+                if (activeScroll instanceof EyesAppiumElement) {
+                    position = (int) (activeScroll.getLocation().getY() * eyesDriver.getDevicePixelRatio());
+                } else {
+                    position = activeScroll.getLocation().getY();
+                }
+                this.contentSize.top = position;
+                this.contentSize.height = activeScroll.getRect().getHeight();
             }
         } catch (NoSuchElementException e) {
             GeneralUtils.logExceptionStackTrace(logger, Stage.CHECK, e);
@@ -385,34 +425,42 @@ public class AndroidScrollPositionProvider extends AppiumScrollPositionProvider 
 
     @Override
     protected WebElement getFirstScrollableView() {
-        WebElement scrollableView = EyesAppiumUtils.getFirstScrollableView(driver);
-        if (scrollableView.getAttribute("className").equals("android.widget.HorizontalScrollView")) {
-            List<MobileElement> list = driver.findElements(By.xpath(EyesAppiumUtils.SCROLLVIEW_XPATH));
-            for (WebElement element : list) {
-                if (element.getAttribute("className").equals("android.widget.HorizontalScrollView")) {
-                    continue;
+        if (cachedScrollableView == null) {
+            WebElement scrollableView;
+            if (scrollRootElement != null) {
+                scrollableView = driver.findElement(MobileBy.id(scrollRootElement.getAttribute("resourceId")));
+            } else {
+                scrollableView = EyesAppiumUtils.getFirstScrollableView(driver);
+                if (scrollableView.getAttribute("className").equals("android.widget.HorizontalScrollView")) {
+                    List<MobileElement> list = driver.findElements(By.xpath(EyesAppiumUtils.SCROLLVIEW_XPATH));
+                    for (WebElement element : list) {
+                        if (element.getAttribute("className").equals("android.widget.HorizontalScrollView")) {
+                            continue;
+                        }
+                        List<MobileElement> child = scrollableView.findElements(By.xpath(EyesAppiumUtils.SCROLLVIEW_XPATH));
+                        scrollableView = child.isEmpty() ? element : child.get(0);
+                    }
                 }
-                List<MobileElement> child = scrollableView.findElements(By.xpath(EyesAppiumUtils.SCROLLVIEW_XPATH));
-                return child.isEmpty() ? element : child.get(0);
             }
+            cachedScrollableView = scrollableView;
         }
-        return scrollableView;
+        return cachedScrollableView;
     }
 
-    private String getScrollableContentSize(String resourceId) {
+    public String getScrollableContentSize(String resourceId) {
         String scrollableContentSize = "";
         String[] version = EyesAppiumUtils.getHelperLibraryVersion(eyesDriver, logger).split("\\.");
         MobileElement hiddenElement;
         if (version.length == 3 &&
                 Integer.parseInt(version[0]) >= 1 &&
-                Integer.parseInt(version[1]) >= 3 &&
-                Integer.parseInt(version[2]) >= 1) {
+                Integer.parseInt(version[1]) >= 3) {
             hiddenElement = ((AndroidDriver<AndroidElement>) driver).findElement(MobileBy.AndroidUIAutomator("new UiSelector().description(\"EyesAppiumHelperEDT\")"));
             if (hiddenElement != null) {
                 String elementId = resourceId.split("/")[1];
                 hiddenElement.setValue("offset;"+elementId+";0;0;0");
                 hiddenElement.click();
                 scrollableContentSize = hiddenElement.getText();
+                hiddenElement.clear();
             }
         } else {
             hiddenElement = ((AndroidDriver<AndroidElement>) driver).findElement(MobileBy.AndroidUIAutomator("new UiSelector().description(\"EyesAppiumHelper\")"));