+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/EyesAppiumUtils.java
@@ -76,6 +76,7 @@ public class EyesAppiumUtils {
             contentSize.top = element.getLocation().getY();
             contentSize.left = element.getLocation().getX();
             contentSize.scrollableOffset = element.getSize().getHeight();
+            contentSize.setDriver(driver);
         }
         return contentSize;
     }